Data type was changed

+++ src/main/java/com/zontik/groshiky/service/AccountService.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 public class AccountService implements IAccountService{
 
     private final AccountRepository accountRepository;
-    private final static float DEFAULT_BALANCE = 0f;
+    private final static double DEFAULT_BALANCE = 0d;
 
     @Autowired
     public AccountService(AccountRepository accountRepository) {
